requirements.txt
# Flask and CORS
Flask==3.1.0
flask-cors==5.0.0
Werkzeug==3.1.3

# LangChain and AI
langchain==0.3.16
langchain-community==0.3.15
langchain-core==0.3.28
langchain-ollama==0.2.2
langchain-text-splitters==0.3.4

# Vector Store
faiss-cpu==1.9.0.post1

# Document Processing
docx2txt==0.8

# Environment and Configuration
python-dotenv==1.0.1

# Validation
pydantic==2.10.5
pydantic-core==2.27.2

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-flask==1.3.0

# Development Tools
black==24.10.0
flake8==7.1.1
